+++ test/smokes/cppcheck/expectations.rb
@@ -492,3 +492,243 @@ s.add_test(
   ],
   analyzer: { name: "Cppcheck", version: "2.0" }
 )
+
+s.add_test(
+  "bug_hunting_target",
+  type: "success",
+  issues: [
+    {
+      id: "bughuntingDivByZero",
+      path: "src/bar/cpp_sample.c++",
+      location: { start_line: 11 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZero",
+      path: "src/cpp_sample_2.c++",
+      location: { start_line: 9 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/bar/c_sample.c",
+      location: { start_line: 18 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/bar/c_sample.c",
+      location: { start_line: 20 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/bar/cpp_sample.c++",
+      location: { start_line: 24 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/cpp_sample_2.c++",
+      location: { start_line: 22 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/bar/c_sample.c",
+      location: { start_line: 16 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/bar/c_sample.c",
+      location: { start_line: 19 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/bar/cpp_sample.c++",
+      location: { start_line: 22 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/bar/cpp_sample.c++",
+      location: { start_line: 25 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_2.c++",
+      location: { start_line: 20 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_2.c++",
+      location: { start_line: 23 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    }
+  ],
+  analyzer: { name: "Cppcheck", version: "2.0" }
+)
+
+s.add_test(
+  "bug_hunting_project",
+  type: "success",
+  issues: [
+    {
+      id: "bughuntingDivByZero",
+      path: "src/cpp_sample_1.c++",
+      location: { start_line: 7 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZero",
+      path: "src/cpp_sample_3.c++",
+      location: { start_line: 11 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/cpp_sample_1.c++",
+      location: { start_line: 20 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "src/cpp_sample_3.c++",
+      location: { start_line: 24 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_1.c++",
+      location: { start_line: 18 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_1.c++",
+      location: { start_line: 21 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_3.c++",
+      location: { start_line: 22 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "src/cpp_sample_3.c++",
+      location: { start_line: 25 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    }
+  ],
+  analyzer: { name: "Cppcheck", version: "2.0" }
+)
+
+s.add_test(
+  "bug_hunting_addon",
+  type: "success",
+  issues: [
+    {
+      id: "bughuntingDivByZeroFloat",
+      path: "c_sample.c",
+      location: { start_line: 12 },
+      message: "There is division, cannot determine that there can't be a division by zero.",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "369", location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "misra-c2012-10.4",
+      path: "c_sample.c",
+      location: { start_line: 12 },
+      message: "misra violation (use --rule-texts=<file> to get proper output)",
+      links: [],
+      object: { severity: "style", verbose: nil, inconclusive: false, cwe: nil, location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "misra-c2012-12.3",
+      path: "c_sample.c",
+      location: { start_line: 5 },
+      message: "misra violation (use --rule-texts=<file> to get proper output)",
+      links: [],
+      object: { severity: "style", verbose: nil, inconclusive: false, cwe: nil, location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "uninitvar",
+      path: "c_sample.c",
+      location: { start_line: 10 },
+      message: "Uninitialized variable: dz",
+      links: [],
+      object: { severity: "error", verbose: nil, inconclusive: false, cwe: "457", location_info: nil },
+      git_blame_info: nil
+    }
+  ],
+  analyzer: { name: "Cppcheck", version: "2.0" }
+)